Create fakeLimeLightTest.py
Again used this repo as storage for a python file used for the beak squad. It's a small refactoring of the current LimeLightInterpreter.java and dependencies to independently test the geometry for an off center lime-light. It seems to work, though somehow in the py there are some numerical instabilities that prove a little too distracting to say so for sure